+++ src/utils/index.ts
@@ -1,5 +1,6 @@
 export * from "./async.js";
 export * from "./format.js";
+export * from "./iterable-text.js";
 export * from "./iterable.js";
 export * from "./literal.js";
 export * from "./load.js";
@@ -7,6 +8,5 @@ export * from "./name.js";
 export * from "./package.js";
 export * from "./pointer.js";
 export * from "./root.js";
-export * from "./source-code.js";
 export * from "./status-code.js";
 export * from "./url.js";
